+++ odoo/addons/test_base/__manifest__.py
@@ -1,0 +1,8 @@
+{
+    'name': 'Test Base',
+    'category': 'Hidden/Tests',
+    'description': """Tests base with specific testing ressources (models, views, etc).""",
+    'depends': ['base', 'test_orm'],
+    'author': 'Odoo S.A.',
+    'license': 'LGPL-3',
+}
