+++ odoo/addons/test_base/tests/__init__.py
@@ -1,0 +1,6 @@
+from . import test_ir_actions
+from . import test_ir_model
+from . import test_ir_qweb_fields
+from . import test_ir_rule
+from . import test_ir_ui_view
+from . import test_res_config
