+++ odoo/addons/test_base/tests/test_res_config.py
