+++ odoo/addons/test_base/tests/test_vectors/image
