+++ odoo/addons/test_orm/tests/__init__.py
@@ -2,21 +2,18 @@ from . import (
     test_access_feedback,
     test_access_monetary_related,
     test_acl,
-    test_action_bindings,
     test_all_models,
     test_api,
     test_autovacuum,
     test_cache,
     test_check_access,
     test_company_checks,
-    test_config_settings,
     test_convert,
     test_create,
     test_convert_env,
     test_domain,
     test_domain_expression,
     test_fields,
-    test_html_converter,
     test_indexed_translation,
     test_inherits,
     test_ir_rules,
@@ -30,7 +27,6 @@ from . import (
     test_performance,
     test_properties,
     test_query,
-    test_qweb_float,
     test_read_group_override,
     test_read_group_private,
     test_read_grouping_sets,
@@ -43,5 +39,4 @@ from . import (
     test_timeit,
     test_transactions,
     # test_uninstall  # Cannot be handled by the runbot. Makes Odoo loop forever when done during an installation.
-    test_views,
 )
